requirements.txt
Flask==3.1.0
gunicorn
psycopg2-binary
beautifulsoup4==4.12.3
Flask-Caching
Flask-Compress